chore rename mutation in hooks
