+++ app/products/page-components/formField/ThumbnailUploader.tsx
@@ -3,7 +3,7 @@
 import { useRef } from 'react';
 import { Controller, Control } from 'react-hook-form';
 import Image from 'next/image';
-import useUploadImage from '@/hooks/api/useUpload';
+import useUploadImage from '@/hooks/mutation/useUpload';
 
 import {
   validateImageFile,
